+++ tests/test_crawl/test_crawl.py
@@ -704,3 +704,27 @@ async def test_wrong_url_in_catalog(
     else:
         check = await Check.get_by_resource_id(RESOURCE_ID)
         assert check["status"] == 404
+
+
+@pytest.mark.parametrize(
+    "check_duration",
+    [
+        config.STUCK_THRESHOLD_SECONDS // 2,
+        config.STUCK_THRESHOLD_SECONDS * 2,
+    ],
+)
+async def test_reset_statuses(fake_check, db, setup_catalog, check_duration):
+    """Reset the status of a resource stuck for a while"""
+    await fake_check(
+        resource_id=RESOURCE_ID, created_at=datetime.now() - timedelta(seconds=check_duration)
+    )
+    status = "ANALYSING_CSV"
+    await db.execute(f"UPDATE catalog SET status = '{status}' WHERE resource_id = $1", RESOURCE_ID)
+    row = await db.fetchrow("SELECT status FROM catalog WHERE resource_id = $1", RESOURCE_ID)
+    assert row["status"] == status
+    await Resource.clean_up_statuses()
+    row = await db.fetchrow("SELECT status FROM catalog WHERE resource_id = $1", RESOURCE_ID)
+    if check_duration > config.STUCK_THRESHOLD_SECONDS:
+        assert row["status"] is None
+    else:
+        assert row["status"] == status