+++ udata_hydra/cli.py
@@ -116,6 +116,8 @@ async def load_catalog(
                     else None,
                 )
         log.info("Resources catalog successfully upserted into DB.")
+        await Resource.clean_up_statuses()
+        log.info("Stuck statuses sucessfully reset to null.")
     except Exception as e:
         raise e
     finally: