+++ udata_hydra/db/resource.py
@@ -1,3 +1,5 @@
+from datetime import datetime, timedelta, timezone
+
 from asyncpg import Record
 
 from udata_hydra import config, context
@@ -127,3 +129,28 @@ class Resource:
                 "(catalog.status IS NULL OR catalog.status = 'BACKOFF')",
             ]
         )
+
+    @staticmethod
+    async def get_stuck_resources() -> list[str]:
+        """Some resources end up being stuck in a not null status forever,
+        we want to get them back on track.
+        This returns all resource ids of such stuck resources.
+        """
+        threshold = (
+            datetime.now(timezone.utc) - timedelta(seconds=config.STUCK_THRESHOLD_SECONDS)
+        ).strftime("%Y-%m-%d %H:%M:%S")
+        q = f"""SELECT ca.resource_id
+            FROM checks c
+            JOIN catalog ca
+            ON c.id = ca.last_check
+            WHERE ca.status IS NOT NULL AND c.created_at < '{threshold}';"""
+        pool = await context.pool()
+        async with pool.acquire() as connection:
+            rows = await connection.fetch(q)
+        return [str(r["resource_id"]) for r in rows] if rows else []
+
+    @classmethod
+    async def clean_up_statuses(cls):
+        stuck_resources: list[str] = await cls.get_stuck_resources()
+        for rid in stuck_resources:
+            await cls.update(rid, {"status": None})